+++ ACadSharp/Entities/IVertex.cs
@@ -1,0 +1,11 @@
+﻿using CSMath;
+
+namespace ACadSharp.Entities
+{
+	public interface IVertex
+	{
+		IVector Location { get; }
+
+		double Bulge { get; }
+	}
+}
